refactor crypt; change memorycrypt encoding form hex to base64
