+++ debian/changelog
@@ -149,5 +149,6 @@ oidc-agent (2.1.3) UNRELEASED; urgency=medium
   * Fixed a problem with unity OP where access token did not have any scope
   * Fixed strange additional parameters in the auth code exchange request 
   * Fixed superflous error logs when checking if a string is a json object
+  * Changed encoding for memory encryption from hex to base64 
 
  -- Marcus Hardt <hardt@kit.edu>  Thu, 13 Dec 2018 13:57:47 +0100
