+++ src/oidc-agent/lock_state.c
@@ -3,7 +3,7 @@
 
 #include "agent_state.h"
 #include "list/list.h"
-#include "utils/cryptUtils.h"
+#include "utils/crypt/cryptUtils.h"
 #include "utils/memory.h"
 #include "utils/oidc_error.h"
