+++ src/utils/crypt/cryptUtils.h
@@ -15,20 +15,10 @@ char* decryptFileContent(const char* fileContent, const char* password);
 char* decryptLinesList(list_t* lines, const char* password);
 int   crypt_compare(const unsigned char* s1, const unsigned char* s2);
 
-char*        hashPassword(const char* pw);
-int          verify_hashedPassword(const char* pw, const char* hash_str);
-oidc_error_t lockEncrypt(list_t* loaded, const char* password);
-oidc_error_t lockDecrypt(list_t* loaded, const char* password);
+oidc_error_t         lockEncrypt(list_t* loaded, const char* password);
+oidc_error_t         lockDecrypt(list_t* loaded, const char* password);
 struct oidc_account* getAccountFromList(list_t*              loaded_accounts,
                                         struct oidc_account* key);
 void addAccountToList(list_t* loaded_accounts, struct oidc_account* account);
-
-#ifndef secFreeHashed
-#define secFreeHashed(ptr) \
-  do {                     \
-    _secFreeHashed((ptr)); \
-    (ptr) = NULL;          \
-  } while (0)
-#endif  // secFreeHased
 
 #endif  // CRYPT_UTILS_H
